Overhaul (Captcha rework and 2captcha reports)
Captcha was revorked because in 9 of 10 runs it was detected incorrectly. Captcha was sent to service incorrectly so I made script to took a screenshot and then crop it. Not efficient but it get the job done.

Also I added a loop for Captcha check. If it was solve incorrectly loop will begin again and report will be send to Captcha service. If everything is OK loop will be broken and report for successful also will be send.