Exception check added
Added ENV to check if Exception check need to be sended. Cause most of the time exception is raised by Captcha service errors which is not related to user or his server. So most of the time those message is not needed.
But in case of Debug you can enable those in ENVs